# docs/node_swagger.yaml
definitions:
...
  api.Events:
    properties:
      address:
        type: string
      data:
        items:
          type: string
        type: array
      identifier:
        type: string
      topics:
        items:
          type: string
        type: array
    type: object
  api.Logs:
    properties:
      address:
        type: string
      contractId:
        type: integer
      events:
        items:
          $ref: '#/definitions/api.Events'
        type: array
    type: object
...
  api.Transaction:
    properties:
      Block:
        type: integer
      GasLimit:
        type: integer
      GasMultiplier:
        type: integer
      RawData:
        $ref: '#/definitions/transaction.Transaction_Raw'
      Receipts:
        items:
          $ref: '#/definitions/transaction.Transaction_Receipt'
        type: array
      Result:
        $ref: '#/definitions/transaction.Transaction_TXResult'
      ResultCode:
        $ref: '#/definitions/transaction.Transaction_TXResultCode'
      Signature:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      hash:
        type: string
      logs:
        $ref: '#/definitions/api.Logs'
      signature:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
...
  data.Event:
    properties:
      address:
        type: string
      data:
        items:
          type: string
        type: array
      identifier:
        type: string
      order:
        type: integer
      topics:
        items:
          type: string
        type: array
    type: object
  data.KDAFee:
    properties:
      amount:
        type: integer
      kda:
        type: string
    type: object
  data.Logs:
    properties:
      address:
        type: string
      contractId:
        type: integer
      events:
        items:
          $ref: '#/definitions/data.Event'
        type: array
      timestamp:
        $ref: '#/definitions/time.Duration'
    type: object
...
  data.TXContract:
    properties:
      parameter: {}
      type:
        $ref: '#/definitions/transaction.TXContract_ContractType'
      typeString:
        type: string
    type: object
  data.Transaction:
    properties:
      bandwidthFee:
        type: integer
      blockNum:
        type: integer
      chainID:
        type: string
      contract:
        items:
          $ref: '#/definitions/data.TXContract'
        type: array
      data:
        items:
          type: string
        type: array
      hasLogs:
        type: boolean
      hasOperations:
        type: boolean
      hash:
        type: string
      kAppFee:
        type: integer
      kdaFee:
        $ref: '#/definitions/data.KDAFee'
      logs:
        $ref: '#/definitions/data.Logs'
      nonce:
        type: integer
      permissionID:
        type: integer
      receipts:
        items:
          additionalProperties: true
          type: object
        type: array
      resultCode:
        type: string
      searchOrder:
        type: integer
      sender:
        type: string
      signature:
        items:
          type: string
        type: array
      status:
        type: string
      timestamp:
        $ref: '#/definitions/time.Duration'
      version:
        type: integer
    type: object
...
  models.BroadcastTXResponse:
    properties:
      txCount:
        type: integer
      txHash:
        type: string
      txsHashes:
        items:
          type: string
        type: array
    type: object
...
  models.SendTXResponse:
    properties:
      result: {}
      txHash:
        type: string
    type: object
  models.SimulateTransactionResponse:
    properties:
      bandwidthFee:
        type: integer
      kappFee:
        type: integer
    type: object
  models.TransactionPoolItem:
    properties:
      hash:
        type: string
      nonce:
        type: integer
      sender:
        type: string
    type: object
  models.TransactionPoolResponse:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      sender:
        type: string
      total:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/models.TransactionPoolItem'
        type: array
    type: object
...
  shared.GenericAPIResponse:
    properties:
      code:
        $ref: '#/definitions/shared.ReturnCode'
      data: {}
      error:
        type: string
    type: object
  shared.ReturnCode:
    enum:
    - successful
    - internal_issue
    - bad_request
    - system_busy
    type: string
    x-enum-varnames:
    - ReturnCodeSuccess
    - ReturnCodeInternalError
    - ReturnCodeRequestError
    - ReturnCodeSystemBusy
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
  transaction.BroadcastTXRequest:
    properties:
      tx:
        $ref: '#/definitions/transaction.Transaction'
      txs:
        items:
          $ref: '#/definitions/transaction.Transaction'
        type: array
    type: object
  transaction.FeesResponse:
    properties:
      bandwidthFee:
        type: integer
      gasEstimated:
        type: integer
      gasMultiplier:
        type: integer
      kAppFee:
        type: integer
      kdaFee:
        $ref: '#/definitions/transaction.Transaction_KDAFee'
      returnMessage:
        type: string
      safetyMargin:
        type: integer
    type: object
  transaction.TXContract:
    properties:
      Parameter:
        $ref: '#/definitions/anypb.Any'
      Type:
        $ref: '#/definitions/transaction.TXContract_ContractType'
    type: object
...
  transaction.Transaction:
    properties:
      Block:
        type: integer
      GasLimit:
        type: integer
      GasMultiplier:
        type: integer
      RawData:
        $ref: '#/definitions/transaction.Transaction_Raw'
      Receipts:
        items:
          $ref: '#/definitions/transaction.Transaction_Receipt'
        type: array
      Result:
        $ref: '#/definitions/transaction.Transaction_TXResult'
      ResultCode:
        $ref: '#/definitions/transaction.Transaction_TXResultCode'
      Signature:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
    type: object
  transaction.Transaction_KDAFee:
    properties:
      Amount:
        description: 'TODO: allow spread'
        type: integer
      KDA:
        items:
          type: integer
        type: array
    type: object
  transaction.Transaction_Raw:
    properties:
      BandwidthFee:
        type: integer
      ChainID:
        items:
          type: integer
        type: array
      Contract:
        items:
          $ref: '#/definitions/transaction.TXContract'
        type: array
      Data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      KAppFee:
        type: integer
      KDAFee:
        $ref: '#/definitions/transaction.Transaction_KDAFee'
      Nonce:
        type: integer
      PermissionID:
        type: integer
      Sender:
        items:
          type: integer
        type: array
      Version:
        type: integer
    type: object
  transaction.Transaction_Receipt:
    properties:
      Data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
    type: object
  transaction.Transaction_TXResult:
    enum:
    - 0
    - 1
    format: int32
    type: integer
    x-enum-varnames:
    - Transaction_SUCCESS
    - Transaction_FAILED
  transaction.Transaction_TXResultCode:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    - 19
    - 20
    - 21
    - 22
    - 23
    - 24
    - 25
    - 26
    - 27
    - 28
    - 29
    - 30
    - 31
    - 32
    - 33
    - 34
    - 35
    - 36
    - 37
    - 38
    - 39
    - 40
    - 41
    - 42
    - 43
    - 44
    - 45
    - 46
    - 47
    - 48
    - 49
    - 50
    - 51
    - 52
    - 53
    - 54
    - 55
    - 56
    - 57
    - 58
    - 59
    - 60
    - 61
    - 62
    - 63
    - 64
    - 65
    - 99
    format: int32
    type: integer
    x-enum-varnames:
    - Transaction_Ok
    - Transaction_OutOfFunds
    - Transaction_AccountError
    - Transaction_AssetError
    - Transaction_ContractInvalid
    - Transaction_ContractNotFound
    - Transaction_FeeInvalid
    - Transaction_ParameterInvalid
    - Transaction_APRInvalid
    - Transaction_AssetIDInvalid
    - Transaction_AssetTypeInvalid
    - Transaction_AssetCantBeMinted
    - Transaction_AssetCantBeBurned
    - Transaction_AssetCantBePaused
    - Transaction_AssetCantBeDelegated
    - Transaction_AssetOwnerCantBeChanged
    - Transaction_AccountNotOwner
    - Transaction_CommissionTooHigh
    - Transaction_DelegationAmountInvalid
    - Transaction_ProposalNotActive
    - Transaction_ValueInvalid
    - Transaction_AmountInvalid
    - Transaction_BucketIDInvalid
    - Transaction_KeyConflict
    - Transaction_MaxDelegationAmount
    - Transaction_InvalidPeerKey
    - Transaction_MinKFIStakedUnreached
    - Transaction_MaxSupplyExceeded
    - Transaction_SaveAccountError
    - Transaction_LoadAccountError
    - Transaction_SameAccountError
    - Transaction_AssetPaused
    - Transaction_DeletegateError
    - Transaction_WithdrawNotAvailable
    - Transaction_ErrOverflow
    - Transaction_SetStakingErr
    - Transaction_SetMarketOrderErr
    - Transaction_BalanceError
    - Transaction_KAPPError
    - Transaction_UnfreezeError
    - Transaction_UndelegateError
    - Transaction_WithdrawError
    - Transaction_ClaimError
    - Transaction_BucketsExceeded
    - Transaction_AssetCantBeWiped
    - Transaction_AssetCantAddRoles
    - Transaction_FreezeError
    - Transaction_ITONotActive
    - Transaction_NFTMintStopped
    - Transaction_RoyaltiesChangeStopped
    - Transaction_ITOKAPPError
    - Transaction_ITOWhiteListError
    - Transaction_NFTMetadataChangeStopped
    - Transaction_AlreadyExists
    - Transaction_IteratorLimitReached
    - Transaction_VMFunctionNotFound
    - Transaction_VMFunctionWrongSignature
    - Transaction_VMUserError
    - Transaction_VMOutOfGas
    - Transaction_VMAccountCollision
    - Transaction_VMCallStackOverFlow
    - Transaction_VMExecutionPanicked
    - Transaction_VMExecutionFailed
    - Transaction_VMUpgradeFailed
    - Transaction_VMSimulateFailed
    - Transaction_KDATransferNotAllowed
    - Transaction_Fail
  transaction.sendTXRequest:
    properties:
      contract:
        items:
          type: integer
        type: array
      contracts:
        items:
          items:
            type: integer
          type: array
        type: array
      data:
        items:
          items:
            format: int32
            type: integer
          type: array
        type: array
      kdaFee:
        type: string
      nonce:
        type: integer
      permID:
        type: integer
      sender:
        type: string
      type:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample of the KleverChain Node Server.
  title: KleverChain Node
  version: "1.0"
paths:
...
  /transaction/{txhash}:
    get:
      consumes:
      - application/json
      parameters:
      - description: txhash
        in: path
        name: txhash
        required: true
        type: string
      - description: include execution results (default false)
        in: query
        name: withResults
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    transaction:
                      $ref: '#/definitions/api.Transaction'
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: some internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns transaction details for a given txhash
      tags:
      - Transaction
  /transaction/broadcast:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: BroadcastTXRequest
        required: true
        schema:
          $ref: '#/definitions/transaction.BroadcastTXRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.BroadcastTXResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: broadcast the TX
      tags:
      - Transaction
  /transaction/decode:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/transaction.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    tx:
                      $ref: '#/definitions/data.Transaction'
                  type: object
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: decode a given transaction
      tags:
      - Transaction
  /transaction/estimate-fee:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/transaction.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/transaction.FeesResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: some internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns fee details for a given transaction
      tags:
      - Transaction
  /transaction/pool:
    get:
      consumes:
      - application/json
      parameters:
      - description: sender (optional)
        in: query
        name: sender
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.TransactionPoolResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: some internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns transaction memory pool
      tags:
      - Transaction
  /transaction/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: sendTXRequest
        required: true
        schema:
          $ref: '#/definitions/transaction.sendTXRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SendTXResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns send transaction
      tags:
      - Transaction
  /transaction/simulate:
    post:
      consumes:
      - application/json
      parameters:
      - description: body data
        in: body
        name: sendTXRequest
        required: true
        schema:
          $ref: '#/definitions/transaction.sendTXRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SimulateTransactionResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: simulate transaction gas consumption
      tags:
      - Transaction
...
swagger: "2.0"